Utilities/requirements.txt
python-dateutil>=2.8.2
python-dotenv>=0.21.0
ibm-cos-sdk>=2.13.0
ibm-cloud-sdk-core>=3.18.0
ibm-platform-services==0.33.1
ibm-vpc>=0.16.0
numpy>=1.24.2
pandas==1.5.3
python-dateutil>=2.8.2
XlsxWriter==3.0.2
python-dotenv>=0.21.0
pytz>=2022.7.1
pyyaml>=6.0
openpyxl>=3.1.1
